Refactor CarouselButtons
